+++ app/assets/javascripts/views/project_view.js
@@ -1,4 +1,3 @@
-import AttachmentOptions from 'models/attachmentOptions'
 import StoryAttachment from 'components/story/StoryAttachment';
 var StoryView = require('./story_view');
 var IterationView = require('./iteration_view');
@@ -26,14 +25,6 @@ module.exports = Backbone.View.extend({
     this.$loadingSpin.show();
     this.model.stories.fetch({success: this.addAll});
 
-    const attachmentOptions = new AttachmentOptions({
-      refreshCallback: (options) => {
-        this.attachmentOptions = options;
-        this.trigger('attachmentOptions', options);
-      }
-    });
-
-    attachmentOptions.fetch();
   },
 
   prepareColumns: function() {
@@ -97,12 +88,7 @@ module.exports = Backbone.View.extend({
     }
     var view = new StoryView({
       model: story,
-      attachmentOptions: this.attachmentOptions
     }).render();
-
-    view.listenTo(this, 'attachmentOptions', (options) => {
-      view.trigger('attachmentOptions', options);
-    });
 
     this.appendViewToColumn(view, column);
     view.setFocus();