+++ config/initializers/cloudinary.rb
@@ -1,0 +1,4 @@
+Cloudinary.config do |config|
+  config.cloud_name = ENV['CLOUD_NAME']
+  config.cdn_subdomain = true
+end