+++ config/initializers/cors.rb
@@ -1,6 +1,8 @@
 Rails.application.config.middleware.insert_before 0, "Rack::Cors" do
   allow do
     origins '*'
-    resource '*', headers: :any, methods: [:get, :post, :options]
+    resource '*', 
+      headers: :any, 
+      methods: [:get, :post, :options]
   end
 end