+++ config/webpack/production.js
@@ -14,9 +14,6 @@ module.exports = merge(sharedConfig, {
 
   plugins: [
     new webpack.NoErrorsPlugin(),
-    new webpack.DefinePlugin({
-      'process.env': { NODE_ENV: JSON.stringify('production') }
-    }),
     new webpack.optimize.DedupePlugin(),
     new webpack.optimize.OccurrenceOrderPlugin(),
     new webpack.optimize.UglifyJsPlugin({